refactor: app-tools builder shared
