+++ .changeset/tricky-coins-warn.md
@@ -1,0 +1,6 @@
+---
+'@modern-js/app-tools': patch
+---
+
+refactor: app-tools builder
+refactor: 重构 app-tools builder
