+++ packages/solutions/app-tools/src/builder/shared/builderPlugins/index.ts
@@ -1,0 +1,3 @@
+export * from './adapterBasic';
+export * from './adapterHtml';
+export * from './adapterSSR';
