+++ packages/solutions/app-tools/src/builder/shared/index.ts
@@ -1,3 +1,3 @@
 export * from './createCopyPattern';
 export * from './types';
-export * from './builderPlugins/adapterModern';
+export * from './builderPlugins';
